Move storage mod into MutinyBrowserStorage
