+++ node-manager/src/lib.rs
@@ -6,7 +6,6 @@ mod error;
 mod localstorage;
 mod nodemanager;
 mod seedgen;
-mod storage;
 mod utils;
 
 use cfg_if::cfg_if;
